+++ src/Indice.Features.Identity.Core/DeviceAuthentication/Validation/CompleteRegistrationRequestValidator.cs
@@ -144,7 +144,7 @@ internal class CompleteRegistrationRequestValidator : RequestValidatorBase<Compl
                 .Create<User>()
                 .VerifyAsync(user, parameters.Get(RegistrationRequestParameters.OtpCode)!, Constants.DeviceAuthenticationOtpPurpose(userId, authorizationCode.DeviceId!, authorizationCode.InteractionMode));
             if (!totpResult.Success) {
-                return Error(totpResult.Error);
+                return Error(totpResult.Error!);
             }
         }
         // Finally return result.