+++ src/Indice.Features.Identity.UI/IdentityUIOptions.cs
@@ -23,7 +23,7 @@ public class IdentityUIOptions
     /// </item>
     /// </list>
     /// </remarks>
-    public string HomePageSlogan { get; set; } = "Welcome to our Digital Services <strong>Portal</strong> of {0}";
+    public string HomePageSlogan { get; set; } = "Welcome to the {0} Digital Services <strong>Portal</strong>";
     /// <summary>An absolute URL to the <strong>terms and conditions</strong> web page. Use it when this page is located to (or shared with) an external website.</summary>
     /// <remarks>If left null the <strong>./legal/terms.md</strong> will be used. If populated it will do a redirect to this URL</remarks>
     public string? TermsUrl { get; set; }